replace sysout with logger

+++ src/main/java/com/hitema/dao/DAO.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface DAO<T> {
     List<T> findAll();
-    T find(int id);
+    T find(Long id);
     void save(T object);
     void update(T object);
     void delete(T object);
